+++ RELEASES.rst
@@ -1846,7 +1846,8 @@ Planned
 * Add support for ES6 String.prototype.codePointAt(), String.fromCodePoint(),
   and String.prototype.repeat() (GH-1043, GH-1049, GH-1050)
 
-* Add support for ES6 Math.hypot() (GH-1069)
+* Add support for ES6 Math.hypot(), Math.cbrt(), Math.log2(), Math.log10(),
+  Math.trunc() (GH-1069, GH-1093)
 
 * Add support for ES6 Object.assign() (GH-1064)